+++ vivi/commands/count.ts
@@ -1,9 +1,9 @@
 import { replyToInteraction } from '../../src/command-handler';
-import { getPromptRegexDisplayText } from '../../src/emoji-renderer';
 import { CommandInteraction, SlashCommandBuilder } from 'discord.js';
 import { formatNumber } from '../../src/utils';
 
 import { cleanWord, solvePromptWithTimeout, getPromptRegexFromPromptSearch } from '../../src/dictionary/dictionary';
+import { getPromptRegexDisplayText } from '../../src/regex';
 
 export const data = new SlashCommandBuilder()
   .setName('count')
